+++ discord/ui/__init__.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from .button import ButtonStyle, Button
 from .select_menu import SelectMenu, SelectOption
-from .models import BaseComponent, DisableableComponent, ComponentHolder, InteractionComponent
+from .models import BaseComponent, DisableableComponent, ComponentHolder, InteractionComponent, InteractableComponent, LayoutComponent
 from .action_row import ActionRow, MessageComponents
 from .modal import Modal
 from .input_text import InputText
@@ -12,6 +12,8 @@ __all__ = (
     "ButtonStyle",
     "SelectMenu",
     "BaseComponent",
+    "InteractableComponent",
+    "LayoutComponent",
     "ActionRow",
     "Button",
     "SelectOption",